+++ RDBqc/NEWS.md
@@ -1,5 +1,7 @@
 # RDBqc 0.0.17.12
  * Improvement of outputs of check_ldf_TaskVII.2 function
+ * Improvement of outputs of RCG_check_loc
+ * Improvement of outputs of RCG_summarize_ind_meas
 # RDBqc 0.0.17.10
  * Improvement of outputs of MEDBS_check_missing_years function
  * Improvement of outputs of FDI_AER_land_landvalue function